+++ app/Filament/Resources/BrandAssetResource/Pages/ViewBrandAsset.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\BrandAssetResource\Pages;
+
+use App\Filament\Resources\BrandAssetResource;
+use Filament\Resources\Pages\ViewRecord;
+
+class ViewBrandAsset extends ViewRecord
+{
+    protected static string $resource = BrandAssetResource::class;
+}